Update README.md
I feel like I'm understanding this better.  I will create a new repository on GitHub to experiment and try things. I'm about to hit the green "Commit changes" button, which will only change my kcavera-patch-1 branch.  It will not change the original hello-world file. This is my current theory so let's put it to the test!